+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryClientFactoryConfig.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 
 public class BigQueryClientFactoryConfig implements BigQueryConfig {
 
+  private final Optional<String> accessTokenProviderFQCN;
   private final Optional<String> credentialsKey;
   private final Optional<String> credentialsFile;
   private final Optional<String> accessToken;
@@ -24,6 +25,7 @@ public class BigQueryClientFactoryConfig implements BigQueryConfig {
   private final ImmutableMap<String, String> bigQueryJobLabels;
 
   BigQueryClientFactoryConfig(BigQueryConfig bigQueryConfig) {
+    this.accessTokenProviderFQCN = bigQueryConfig.getAccessTokenProviderFQCN();
     this.credentialsKey = bigQueryConfig.getCredentialsKey();
     this.credentialsFile = bigQueryConfig.getCredentialsFile();
     this.accessToken = bigQueryConfig.getAccessToken();
@@ -40,6 +42,11 @@ public class BigQueryClientFactoryConfig implements BigQueryConfig {
     this.endpoint = bigQueryConfig.getEndpoint();
     this.cacheExpirationTimeInMinutes = bigQueryConfig.getCacheExpirationTimeInMinutes();
     this.bigQueryJobLabels = bigQueryConfig.getBigQueryJobLabels();
+  }
+
+  @Override
+  public Optional<String> getAccessTokenProviderFQCN() {
+    return accessTokenProviderFQCN;
   }
 
   @Override